Add dynamic obstacle avoidance demo
This is a little bit hacky and only implements dynamic obstacl avoidance
to the last-link collision sphere in the self collision model. But, it
seems to work well even still.

Future todos -- collision avoidance to all links, + maybe using the true
relative velocity of each collision sphere? This might not be needed though